requirements.txt
asgiref>=3.6.0
pytz==2021.1
sqlparse==0.4.1
setuptools==69.0.0
drf-spectacular==0.27.2
python-dotenv==1.0.1
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=23.5
djoser>=2.3